+++ spec/fuzz/seed_corpus.rb
@@ -53,6 +53,7 @@ CSS_SEED = [
   "div > p", "section p", "ul li + li", "li ~ li",
   "a[href]", 'a[href="x"]', '[data-i="1"]', '[class~="hot"]', '[lang|="ja"]',
   "li:first-child", "li:last-child", "li:nth-child(2)", "li:nth-of-type(odd)",
+  ":first-of-type", ":only-of-type", ":nth-of-type(2n+1)", "*:last-of-type",
   "p:not(.lead)", ":root", "section:has(p)", "div, span, a",
   "svg circle", "header, footer", "*:empty",
   # complex (combinator) arguments inside :is/:where/:not + :has variants